Rename opentrons.commands -> opentrons.legacy_commands.
