+++ api/src/opentrons/legacy_commands/__init__.py
@@ -1,0 +1,1 @@
+"""Command models from before v5.0, before Protocol Engine."""
