+++ api/src/opentrons/protocol_api/protocol_context.py
@@ -20,9 +20,13 @@ from opentrons.types import Mount, Location, DeckLocation, DeckSlotName, Staging
 from opentrons.legacy_broker import LegacyBroker
 from opentrons.hardware_control import SyncHardwareAPI
 from opentrons.hardware_control.modules.types import MagneticBlockModel
-from opentrons.commands import protocol_commands as cmds, types as cmd_types
-from opentrons.commands.helpers import stringify_labware_movement_command
-from opentrons.commands.publisher import CommandPublisher, publish, publish_context
+from opentrons.legacy_commands import protocol_commands as cmds, types as cmd_types
+from opentrons.legacy_commands.helpers import stringify_labware_movement_command
+from opentrons.legacy_commands.publisher import (
+    CommandPublisher,
+    publish,
+    publish_context,
+)
 from opentrons.protocols.api_support import instrument as instrument_support
 from opentrons.protocols.api_support.deck_type import (
     NoTrashDefinedError,
