+++ api/src/opentrons/protocol_runner/legacy_command_mapper.py
@@ -6,7 +6,7 @@ from typing import Dict, List, Optional, Tuple, Union
 
 from opentrons_shared_data.pipette.dev_types import PipetteNameType
 from opentrons.types import MountType, DeckSlotName, Location
-from opentrons.commands import types as legacy_command_types
+from opentrons.legacy_commands import types as legacy_command_types
 from opentrons.protocol_engine import (
     ProtocolEngineError,
     actions as pe_actions,
