+++ api/tests/opentrons/test_legacy_broker.py
@@ -2,8 +2,8 @@
 
 from typing import List, NamedTuple, cast
 
-from opentrons.commands.types import CommandMessage
-from opentrons.commands.publisher import CommandPublisher, publish
+from opentrons.legacy_commands.types import CommandMessage
+from opentrons.legacy_commands.publisher import CommandPublisher, publish
 
 
 def _my_command(arg1: int, arg2: str = "", arg3: str = "") -> CommandMessage:
